requirements.txt
absl-py==0.9.0
aniso8601 @ file:///tmp/build/80754af9/aniso8601_1597771691111/work
astor==0.8.1
astunparse==1.6.3
attrs @ file:///tmp/build/80754af9/attrs_1598374659300/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.11.8
cffi @ file:///tmp/build/80754af9/cffi_1598370769933/work
chardet==3.0.4
click==7.1.2
cryptography @ file:///tmp/build/80754af9/cryptography_1598892038851/work
docopt==0.6.2
Flask==1.1.2
flask-redis==0.4.0
flask-restx @ file:///home/conda/feedstock_root/build_artifacts/flask-restx_1593008484882/work
gast==0.3.3
google-auth @ file:///tmp/build/80754af9/google-auth_1598987460909/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio @ file:///tmp/build/80754af9/grpcio_1597424474635/work
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
httpserver==1.1.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446406207/work
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1593112171383/work
Keras-Preprocessing==1.1.2
Markdown @ file:///tmp/build/80754af9/markdown_1597433240441/work
MarkupSafe==1.1.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mysql-connector==2.2.9
mysql-connector-python==8.0.22
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pika @ file:///home/conda/feedstock_root/build_artifacts/pika_1589460142612/work
prompt-toolkit==1.0.14
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyInquirer==1.0.3
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis @ file:///tmp/build/80754af9/redis-py_1593435590316/work
regex==2020.7.14
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scipy==1.4.1
six==1.15.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.7.0
tensorflowjs==2.3.0
termcolor==1.1.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
